ESC-236 move controller models into models package
 * this seems reasonable for now but the use case should be reviewed to
   determine if there is a better location for these classes
