+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/controllers/EligibilityController.scala
@@ -22,6 +22,7 @@ import play.api.data.Forms.mapping
 import play.api.mvc._
 import uk.gov.hmrc.eusubsidycompliancefrontend.actions.EscActionBuilders
 import uk.gov.hmrc.eusubsidycompliancefrontend.config.AppConfig
+import uk.gov.hmrc.eusubsidycompliancefrontend.models.FormValues
 import uk.gov.hmrc.eusubsidycompliancefrontend.models.types.EORI
 import uk.gov.hmrc.eusubsidycompliancefrontend.services.{EligibilityJourney, Store}
 import uk.gov.hmrc.eusubsidycompliancefrontend.views.html._
