+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/models/FormValues.scala
@@ -14,6 +14,6 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.eusubsidycompliancefrontend.controllers
+package uk.gov.hmrc.eusubsidycompliancefrontend.models
 
 case class FormValues(value: String)
